pyproject.toml
[tool.poetry]
name = "basicpyvlc"
version = "0.1.0"
description = ""
authors = ["AccelQuasarDragon <138998466+AccelQuasarDragon@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-vlc = "^3.0.20123"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
